enrolled_job/requirements.txt
# requirements.txt

requests==2.31.0
beautifulsoup4==4.12.2
rarfile==4.0
pyunpack==0.3.7
patool==1.12
SQLAlchemy==2.0.20
pandas==2.0.3
python-dotenv==1.0.0
pymysql==1.0.3